+++ src/app/data/contact.ts
@@ -1,0 +1,6 @@
+export interface Contact {
+  id: string;
+  name: string;
+  phone: number;
+  email: string;
+}
